[Analysis API] Compile files by modules in 'KaFirCompilerFacility'
Before this commit, files in 'KaFirCompilerFacility' were compiled one
by one. Such an approach works good enough in simple cases, but if a few
files have to be analyzed together (because of inlines, sophisticated
common-JVM relations, and so on), it immediately led to BE exceptions
of all kinds.

The new approach groups files in chunks, and files in a single chunk
are compiled together. The chunk doesn't necessarily include files from
only a single module – in KMP projects, common and JVM sources may be
compiled together. Also, currently a code fragment and its context file
are compiled in a single chunk to make local declaration IR accessible
by a code fragment.

The new compilation logic now also handles diagnostics coming from
fir2ir and BE runs.

^KT-72313 Fixed