[K/N] Upgrade reflection target in new function reference node
If there is a reference to a local function, it needs to be updated
in LocalDeclarationsLowering. Calls are updated automatically, as they
are now regular calls, but there is also a reference in reflection
target.

By some accident, it worked without updated while we did this in memory,
but serialization and deserialization had no chances.

^KT-72931