JVM: generate super calls correctly
See the comment on `resolveSuperCallOfFakeOverride`.

The main issue with the old implementation was that
`findSuperDeclaration` could call `resolveFakeOverride` on some fake
override in the middle of the override hierarchy, and the generic
signature of that implementation could be way different from the
method that we need to call at the bottom of the hierarchy, resulting in
incorrect behavior or even StackOverflowError.

 #KT-38029 Fixed
 #KT-73800 Fixed