JVM: move "DefaultImpl redirection" detection logic to JvmCachedDeclarations
This will be used in the future to check if a function from dependencies
(which is not lowered because we don't lower IR for dependencies) has a
DefaultImpls redirection, i.e. if it's a fake override in a class with
an implementation in a superinterface compiled in
`-Xjvm-default=disable` mode.