[Raw FIR, LT] Convert annotations without iterating modifier list twice
The reason, annotation conversion was separated from modifier conversion
is that annotation must be converted inside a `withContainerSymbol`
block.
We move the collection of annotation nodes inside the loop where
modifiers are converted, but convert them later.
This will make it easier to collect context parameters which will be
moved inside the modifier list in a following commit.