[Gradle] Raise the KMP incompatibility with Gradle Java plugin to error
Currently, Gradle does not allow to co-exist nicely two ecosystem
plugins. It is even worth with built-in Gradle plugins as they could be
only changed in new Gradle releases.

We are changing the diagnostic severity in Gradle versions 8.7+ where
this whole incompatibility become an error.

Also rename diagnostic to make the name clearer.

^KT-66542 In Progress