[FIR] Make FirErrorProperty extend FirProperty instead of FirVariable
We need to reuse the existing machinery for analyzing
the initializer of FirProperty. In the next commit it will now also
be applied for erroneous destructuring declarations,
once we start saving their initializer.

Remove `when` branches that now become redundant.

Update some visitors that now need to override
visitErrorProperty/transformErrorProperty,
as revealed by the existing failed tests.

KT-72908