IR Scripting: force script plugin application in IDE's eval expr.
Since the script lowering was moved to the scripting plugin, expression
evaluation in IDE is broken for Kotlin scripts, because the IR compiler
plugin extensions are disabled in this scenario.
This fix introduces the flag to enforce plugin extension application in
this scenario, and enables it for the script lowering extension.

#KT-74004 fixed