[CFG] Add explicit enter node reference to finally block exit node
When gathering returned expressions from a SAM lambda, there is
potential for an infinite sequence of CFG nodes because of loops in the
graph. So instead of searching backwards for the enter node of a finally
block, add a direct link to the finally block exit node. This avoids the
potential infinite sequence and is more efficient than searching.

^KT-74242 Fixed