[Test] Correctly register source transformers for FirNativeSerializationTestGenerated
There is an infrastructure that differently registers source transformers
for generated tests for Native and non-Native tests. This is done this way
because the Native test infrastructure has different test lifecycle
compared to the main compiler test infra, and transformers should be
registered before any test is executed.

`FirNativeSerializationTestGenerated` is a Native test, but it uses
the regular test infra, so these transformers should be registered in
the test method instead of test class constructor