[Analysis API] Provide safer heuristic for matching class filter
Some synthetic classes may lack references to an originating 'PsiFile'.
However, now if we can find at least some class with a proper source
that matches the filter, all nearby classes will also match.

The heuristic may fail if some classes have '$' in names. This is rather
uncommon, though, and having a false positive in that case is not that
critical (not only the class should have clashing '$'-including name,
its source needs to also be 'null').