K2: don't erase "alien" callables in raw type substitution scope
Before this commit, any callable member in a raw type substitution scope
was substituted by erasing its signature. In fact, it should be done
only for callable members dependent on a raw type owner,
so the containing class of a callable member
should be the raw type owner itself,
or it should be an inner class of the raw type owner with any depth.
Otherwise, it's nothing to replace, and raw type scope member
should be kept as is.

Fix back KT-24239
Related to KT-74148