[FIR] Refactor FirContextReceiversDeclarationChecker
In preparation for the change to how we parse context parameter lists,
the checker is rewritten to report UNSUPPORTED for any but the
explicitly allowed declaration types (we go from a denylist to an
allowlist approach).
This is require because context parameter lists will become parseable
on more declarations but the language rules will still forbid them.

#KT-51258