[K/JS] Fix bug with per-file granularity and incremental compilation
The bug is quite 'funny' (maybe because I introduced it). So, the problem was the following:
if a user doesn't use the TypeScript definitions generating feature, the incremental compilation will not produce export files
in per-file granularity. This is because I added code that recompiles the export file only when the d.ts file hash is changed.
It means that if there are no d.ts, the hash is always `null`, so the compiler should use the old export files (for the first compilation, it's `null`)

Also, during the work on the bug, I realized that we recompile and rewrite export files with the same content when changing TypeScript definitions.
It was also fixed in this commit.

^KT-70652 Fixed
^KT-71365 Fixed


Merge-request: KT-MR-19728
Merged-by: Artem Kobzar <Artem.Kobzar@jetbrains.com>