[FIR] Reuse the CS when calculating bare types
Fix the inconsistency in `rawCastToStarProjection_Ok.kt`.

Report `UNSAFE_DOWNCAST_WRT_VARIANCE`. It's similar to
`CANNOT_CHECK_FOR_ERASED_DEPRECATION_WARNING`, but is needed for
situations that are not captured by the cast checker: when the RHS bare
type is calculated during resolve. Otherwise, we risk introducing
breaking changes. See `comparisonOfBuiltInTypesUnderOr` for an example.

^KT-7972