[FIR] Reuse `findSubType` and `findSuperType` from `ResultTypeResolver`
Note: approximating to `findSuperType()` during `is`-checks would also
fix `isCheckAndVarianceWithRecursiveBound.kt`, though at the expense of
making the compiler less smart.

- The whole point of `isCastToTargetTypeErased()` was to avoid
  approximation/filtering of the constraints
- Fix the variable to the respective bound
  - This guarantees that we check exactly the bound against the
    corresponding argument of the RHS type (if the CS says there's no
    contradiction, then whatever it infers must be its own upper bound
    in terms of the type variables we're working with)

For example, for `a: Array<out Number>` and `a is Array<Int>`, we know
that `Var(S) <: Number` and want to check that `Var(S)` may not be
anything that is not `Int`. Adding just `Var(S) <: Int` just
"narrows down" `Var(S)` for the CS: it's not a contradiction. If we were
adding a judgment `∀ Var(S) <: Number. Var(S) <: Int`, then we would
get a contradiction, but we don't support such judgments. This commit
fixes the issue by adding `Number <: Var(S)`. We can't just take the
bound and add `Number <: Int` as that would only work if the bound
is non-recursive: otherwise, we'd still (at some point) end up with a
subtyping tha would further constraint `Var(S)` (to express the actual
true bound, we'd have to write an infinite recursive type, which we
can't do currently).

See `isCheckAndVarianceWithRecursiveBound.kt` for more details. This
commit makes sure this test results in an actual contradiction in
the CS without approximating/filtering out the constraints.