KT-74600 Include ExperimentalAtomicApi when serializing builtins
When an annotation class is not resolved,
the current builtin compilation pipeline ignores it.
For common atomic types it led to ExperimentalAtomicApi being missed
on JVM (meaning that users can use experimental atomic types
without an explicit opt-in).
This change fixes the problem by including the
annotation source file during builtin serialization.

^KT-74600

Co-authored-by: vladislavgrecko <vladislav.grechko@jetbrains.com>