Tests: rename K1 directive JVM_ANNOTATIONS -> WITH_KOTLIN_JVM_ANNOTATIONS
In tests-common-new, there's a directive
`ConfigurationDirectives.WITH_KOTLIN_JVM_ANNOTATIONS`
which adds kotlin-annotations-jvm.jar to the classpath.

Using this directive for K2 loadJava tests resolves the issue that
Mutable/ReadOnly annotations were unresolved in these tests. This didn't
led to any failures because errors are not reported in these tests, but
annotations were missing from the FIR dumps.