KT-74641: Bypass std::chrono::system_clock::now() call check
std::chrono::system_clock::now call was introduced previously
as part of GCUnsafeCall-call-chain. The call itself is expected
to be fast and safe, so the thread state transition is not required.

The change adds system_clock::now to the list of "good" functions
whose calls should be ignored by the CallsChecker.

^KT-74641