[Test] Update the `kt54028_cursed.kt` test
This test is supposed to test the compiler behavior with two different
source files with the same name. Since it's impossible to create to real
files with the same path on the disc, the test is rewritten to make it
closer to the real world scenario with `-Xklib-relative-path-base`
CLI argument.