[BE] Support external bytecode providers for inlining
The commit introduces a 'CompiledCodeProvider', a new backend API for
contributing externally available class file data.

The API is needed for the 'KaFirCompilerFacility' which compiles modules
partially for the IDE use-cases (including code evaluation in the JVM
debugger, Compose Live Preview, Bytecode tool window and so on).

Although there is usually a single target file (a file for which the
bytecode should be provided), that file may contain calls of 'inline'
functions from other files or even different modules. It means that
bytecode from other files/modules needs to be somehow provided to the
target file compilation.

Until this commit, the 'KaFirCompilerFacility' manually populated
the 'InlineCache'. Although this worked in simple cases, the cache
implementation is based on SLRU maps, making the setup prone to sporadic
element evictions.

The new API allows passing the class data directly. Together with
solving the original problem, it also makes it possible to pass output
of compiled modules in a whole. E.g., if the IDE can prove the dependent
module sources have not changed, it can reuse class files for more
robust and performant code evaluation.