[Native] Lower bodies from header klibs
When compiling with caches and header klibs, we cannot retrieve inline
function bodies from the klib using the metadata saved in the cache, as
the IR body indices do not match.
We therefore deserialize inline functions for header klibs upfront and
therefore need to lower the bodies the first time they are resolved in
NativeInlineFunctionResolver.


Merge-request: KOTLIN-MR-1321
Merged-by: Johan Bay <jobay@google.com>