[Analysis API] Use compiled classes of 'kotlin-stdlib'
This commit prevents code evaluator from compiling source dependencies
from the 'kotlin-stdlib' module when debugging the Kotlin compiler
project. Instead, the code compiler uses binary class files that are
supposed to exist (as, otherwise, the debug session will not start).

The Kotlin compiler project is special as the Kotlin standard library
is available in sources (all other projects depend on the JAR instead).
Such an unusual setup makes it challenging to evaluate code fragments:

- Evaluated code often includes references to inline functions from
  stdlib, such as map, filter or let. To provide bytecode to the JVM
  inliner, files with the used functions also have to be compiled.
- Stdlib files containing those functions are huge, as they include tens
  of overloads for various receiver types.
- Built-in classes that often require special treatment are also there.
- A huge part of the stdlib is commonized, adding further complexity.

All this contributes to a very low ratio of significant evaluations in
the Kotlin project.

As a fix, the commit replaces compilation of the kotlin-stdlib by
ignoring all files with the 'kotlin' package with binary data of their
containing module. Instead, a provider with the stdlib class output
folder is attached to subsequent chunk compilations.

Theoretically speaking, evaluation results may be invalid if someone
decides to debug the 'kotlin-stdlib' code while changing implementations
of methods. However, this scenario is believed  be very rare or even
non-existent.

^KT-74582 Fixed