JVM: Check for null before unboxing inline class parameter
of suspend lambda during unspilling at the beginning of suspend
lambda.
For normal reference types it works, because nulls are spilled
to continuation if the reference is dead, to prevent memory leaks.
However, if the parameter is inline class, we generate calls to
`unbox-impl`, which, of course, lead to NPE is the parameter is
null.
 #KT-74729 Fixed