[K/N] Introduce `RuntimeInitializer` type in the code generation
Kotlin/Native runtime has a concept of `Initializer` (see the type
declaration in `Runtime.h`).

The code generator generates functions of this type. But the
implementation is convoluted, with many different things called "init"
or "initializer".

This commit encapsulates `Initializer` functions to the new
`RuntimeInitializer` type in the code generator, making the
implementation easier to reason about.

^KT-74764