[JS] Use the single global counter for generating associated object keys
Previously, we'd store the map of associated object keys
in `IrNamerBase`, which is created one per class.
That would mean that if we had two or more
`@AssociatedObjectKey`-annotated annotation classes,
they were assigned the same key of 0, which was wrong.

Instead, we now store the counter of associated object keys in
`JsIrBackendContext`, which means that associated object keys
are always unique.

Note, however, that this solution will not survive incremental
compilation.
This will be fixed in the next commits.

^KT-70132 Fixed