[Build] Configure API/language version 1.7 for Gradle plugins
Also fixes part of deprecation warnings so that project sync passes.
The rest is disabled to be fixed in the next commits
Note that even though we could use language version 1.7, it's preferrable to have it exactly the same as the bundled compiler for Gradle Kotlin scripts for more reliable compatibility.
^KT-72329 In Progress