[K/N] Move creating "init nodes" closer to their usage
The runtime initialization implementation in the Kotlin/Native code
generation operates the concept of "init nodes".
The implementation is split into two parts, one creating init nodes
and the other one gathering them in the right order.

This commit applies a trivial refactoring, moving creation of the init
nodes closer to their grouping and ordering. This makes the code easier
to reason about and to evolve.

^KT-74764