[IR] Consistently use $extension_receiver name for extension receiver
Previously, both dispatch receiver and extension receiver tended to be
named <this>. This created an ambiguity when inspecting IR when a
function had both those receivers, as both had identical names.
Now a distinct name is used for extension receiver.
Also, it brings more consistency from naming such as
<this>, $this or $receiver.

JVM bytecode should be not be affected by this change, as it uses a
special naming for context parameters in codegen.

KT-74275