[Gradle] Depend on published kotlin-stdlib matching the API version
Instead of depending on snapshot kotlin-stdlib from the current revision and limiting it by API version, we are now using the actual stdlib artifact matching the one available at Gradle runtime of the minimal supported version.
That resolves the issue that the compiler used for compilation of Gradle-related modules is outdated for keeping ability to compile targetting older Kotlin language version and may not read the metadata of snapshot kotlin-stdlib.

^KT-61911 Fixed