[FE] Update the project scope before updating KotlinClassFinders on classpath update
The classpath update in `KotlinCoreEnvironment.updateClasspath` consists
of the following steps:
1. Compute the new roots and deduplicate existing ones
2. Add new roots to the `JvmDependenciesIndex` used by the search scope
3. Update the `initialRoots` list or add the new roots to existing class
   finders

If some kotlin class finders were already created with the search scope
from the environment, it's crucial to execute the 2. before the 3.
Otherwise, the newly added roots would be filtered out by scope
in the `JvmPackagePartProvider.addRoots`.

Before this commit the order of steps was 1-3-2, and after it's 1-2-3

^KT-74768 Fixed