[Native] Workaround to support caching of platform libraries on MinGW
MinGW platform libraries are messed up: a lot of random headers tied
together without proper filtering. Fixing them requires a lot of effort.
To mitigate this problem and enable compiler caches on MinGW, we store
caches for MinGW platform libraries as a bitcode: this leaves LLVM DCE
a chance to remove unnecessary declarations. Thus, symbols which could
not be linked for some reason are going to be stripped.