[Native] Enable stdlib caching in tests
MinGW x64 platform libraries are messed up:
* Classification seems random
* Hard to filter actually important headers
* No proper linker flags

Together, these factors make it impossible to properly support compiler
caches for this target.
On the other hand, to speedup compiler tests on CI we don't need a
full-blown cache machinery. So as a compromise, we enable compiler caches
for stdlib in the test infrastructure.