[Analysis API] Add tests for partial body analysis
In this commit, the tests do not yet run partial body analysis.
This is for a purpose – some existing test data differs between the
compiler and LL, and also between normal and reversed analysis.

Changing the order of declaration analysis already produces some effects
on the analysis result. By incorporating the partial body resolution in
an individual commit, such changes can be overseen much easier.

^KT-72357