Allow restarting overridden functions in a final class
Open functions were previously restricted from being restartable because a virtual call made in restart lambda.
This is not an issue for final classes, however, since language guarantees that those classes will not be extended, and that makes restart lambda codegen valid. This change also unblocks some frameworks (like molecule or circuit) that expect final methods to be restartable for optimal performance.

Bug: 377499888
Test: Compose compiler tests