K2: limit applicability of isRelatedToOuterTypeVariable during variable fixation
In PCLA, it's forbidden to fix type variables in case they are
bound to outer type variables with some constraints.
However, this limitation makes no sense for UNTIL_FIRST_LAMBDA mode,
as it's used only for "overload by lambda return type" and in this mode
we have to fix, or at least try to fix, inner type variables,
otherwise the overload just fails.

#KT-74819 Fixed