IR Dump: Workaround for mismatching offsets in FO annotations
Annotations in fake overrides are created by making a deep-copy
of the overridden declaration's annotations. This results in
preserving offsets of original annotations and all the relevant
IR elements (ex: default value expressions) in their copies.

If it happens that the overridden declaration and the container
of the fake override are in different files, then the offsets
of copies are invalid. This can be seen in IR deserialization
tests.

The problem has been described in KT-74938, where it should be
completely fixed. This commit brings just a temporary workaround
to make IR deserialization tests green.

^KT-74456 Fixed