CLI: rename fields for JVM default options back
The field `jvmDefault` was renamed to `jvmDefaultOld` in 39400b350d, and
a new field `jvmDefault` was added in 02418db810, so that it would be
clear which field represents the old experimental argument, and which
field represents the new stable argument.

Unfortunately, in practice this leads to problems with existing JPS
projects (as evidenced by modularized tests) where, because of
KTIJ-7513, the _field names_ (not the string representation of the
compiler arguments) are stored in the Kotlin facet.

So we'll keep using the field `jvmDefault` for the old experimental
argument `-Xjvm-default`, and rename the field corresponding to the new
stable argument `-jvm-default` to `jvmDefaultStable`.