[FIR] Simplify SAM conversion logic
isSubtypeForSamConversion contained some custom logic for checking
the compatibility of some type with an expected function type.
The concreate bug in javaSamTypeMismatch was caused by an unnecessary
call to unwrapToSimpleTypeUsingLowerBound() for comparing the return
types.
The whole custom logic was replaced with a call to
isSubtypeConstraintCompatible.
After the change, we need to account for SAM conversion with function
kind conversion (e.g., suspend conversion).
isSubtypeForSamConversion implicitly accounted for it by comparing
the parameter and return types of the contributed invoke symbols but
ignoring their suspend modifiers.

#KT-74469 Fixed
#KT-52152 Fixed