[Test] Guard hacky compilation of module-info.java in `JvmEnvironmentConfigurator`
On one hand, tests for foreign annotations depend on `module-info.java`
files in dependency modules (and these files are needed to be compiled
to properly be consumed by the kotlin compiler).
On the other hand, these tests declare source dependencies between
modules (both in K1 and K2), so the java compiler is not executed by
regular tests mechanism in some JVM backend facade.

To work around this issue, there is a hack in `JvmEnvironmentConfigurator`
which compiled such files during creation of the compiler configuration.
And this hack started to interfere in some tests which were migrated to
the full compilation pipeline with backend: because of that the module
files in dependencies were duplicated in some case.

It's quite hard to move this hack to `JvmForeignAnnotationsConfigurator`,
as it heavily depends on the order of addition of entries to the content
roots, so it was at least explicitly guarded with check if the test
is actually a test for foreign annotations.

^KT-74856