[K/N] Add `convertBlockPtrToKotlinFunction` intrinsic function
This function is intended to be used for Swift Export. It converts
an Objective-C block pointer (represented as `NativePtr`) to a Kotlin
function object using a statically known function type, following
Objective-C import rules for parameter and return type conversion.
Naturally, it fully reused the Objective-C import implementation for
block pointers.

This intrinsic makes it possible for Swift Export to reuse its existing
type conversions for passing objects through closures: the conversions
between Kotlin and Swift types can be done as usual, through bridge
types, which are then passed through blocks converted with
the intrinsic.

^KT-73878