[AA] Forcefully perform ComposableFunction conversion in stub -> FIR deserializer
Since 2.1 any `@Composable FunctionN` type is serialized to metadata as
`ComposableFunctionN`, which is consistent with how composable functions
are treated in sources (with compose plugin enabled). But there are old
libraries compiled with 2.0 or less, which still have `@Composable FunctionN`
types. To handle such libraries in the CLI compiler plugins are passed
to the library session so they could be applied to deserialized classes.
But it's impossible to do the same in the IDE, because there libraries
don't know anything about source modules they will be used in.

To work around this issue, it was decided to hardcode this type conversion
directly into stub based FIR provider exactly for Compose plugin.

^KTIJ-33020 Fixed