fixup! [FIR] Unify `subject` and `subjectVariable` in when expressions
In FIR smart-casted types from initializers of variables are not saved
to the types of the variables themselves (see`storeVariableReturnType`
in `FirDeclarationsResolveTransformer`). Ensuring the IrVariable's type
is as narrow as that of the initializer is important, for example, for
comparisons with `Double`: if the subject and the branch are both
`Double`s, we may try performing `ieee754` comparison, which will fail
with "Should be primitive or nullable primitive type" if the lhs operand
is not `Double` (which, in the case of when subject, will be `IrGetCall`
to `<when-subject>`).

This change is backed by, for example,
`FirPsiBlackBoxCodegenTestGenerated.testWhenNullableSmartCast`.

Don't update types of all properties. Apparently, there is some unknown
to me invariant about the types of local properties in the presence of
smartcasts in Fir2Ir, because updating the type unconditionally breaks
`compiler/testData/codegen/box/ieee754/when_properIeeeComparisons.kt`:

```
native/native.tests/build/t/bb.src/oa64_FirNatCodBoxTesGen/box/ieee754/when_properIeeeComparisons/main/when_properIeeeComparisons.kt:7:9: error: [IR VALIDATION] after Compiler Phase @UpgradeCallableReferences: unexpected type: expected kotlin.Double, got kotlin.Any
GET_VAR 'val plusZero: kotlin.Double [val] declared in <root>.box' type=kotlin.Any origin=null
  inside TYPE_OP type=kotlin.Boolean origin=INSTANCEOF typeOperand=kotlin.Double
    inside BRANCH
      inside WHEN type=kotlin.Unit origin=IF
        inside BLOCK_BODY
          inside FUN name:box visibility:public modality:FINAL <> () returnType:kotlin.String
            inside FILE fqName:<root> fileName:/Users/Nikolay.Lunyak/Documents/Projects/kotlin-worktrees/kotlin-when-exhaustiveness-improvements/native/native.tests/build/t/bb.src/oa64_FirNatCodBoxTesGen/box/ieee754/when_properIeeeComparisons/main/when_properIeeeComparisons.kt
...
```

^KT-75049 Fixed