[Swift export] sir-compiler-bridges now depend on analysis-api
For historical reasons, sir-compiler-bridges was abstracted from
Analysis API details. While it was a good thing in the beginning of the
project, now this abstraction makes things worse: One have add more and
more "Kotlin modeling" code instead of directly accessing Kotlin origin
which was used to create a SIR node.

So let's fix that and simplify our lives.

^KT-75079