K2: report nullability warning around KT-71718
This commit is a preparation for dropping so-called
TypePreservingVisibilityWrtHack. After this operation
some tests will become red, so now we report warnings about it.

The new warnings are reported here in situations with
a nullable explicit type argument, that is replaced with
a flexible type in master by TypePreservingVisibilityWrtHack.
To do it, we apply a special attribute to such types and then
re-use logic inside FirExpressionJavaNullabilityWarningCheckers.kt
to report type mismatch warnings.

#KT-71718 In Progress