[LL FIR] properly support context parameters
Now they are fully supported and optimized like regular value parameters.
Changes in `analysis/analysis-api/testData/symbols/singleSymbolByPsi/contextParameters`
are expected as now the corresponding FIR declaration is the class
and not the constructor. Both of them have the context parameter
as a real source, which is incorrect but most likely won't be fixed
as context parameters on classes are deprecated.
`analysis/low-level-api-fir/testData/lazyResolve/contextParametersOnClass.kt`
reproduces the same issue.

^KT-73493 Fixed