[K/N] Do not DCE bridges targets
Sometimes because of different optimizations the bridge target gets inlined into
the bridge itself, leading to the target being unreachable in call graph which
then leads to DCE-ing it away. But they are actually used later during the classes
layout construction (before the codegen), so they need to be kept as is.

 #KT-74377 Fixed