[KLIB Tool] Extract size details from KLIB files
Previously, the KLIB tool was only capable of extracting
the size details if a KLIB was represented as a directory.
And if a KLIB was an ZIP archive file, then it just printed
the archive file size without detalization.

Now, with this commit, the KLIB tool is able to show the detalization
for KLIB archives as well as for directories.