[TEST] 2/2 Allow TEST_ONLY language features during testing
KT-75113

The commit fixes a bunch of org.jetbrains.kotlin.konan.test.blackbox.FirNativeCodegenBoxTestGenerated tests,
that got broken by the previous commit. In total, 53 tests were broken
by the previous commit.

The downside of this commit is that it's no longer possible to "test"
the effect of `TEST_ONLY` in our test infra. Well, yes, it's part of the
contract of `TEST_ONLY`, I think it's fine, that's why I have to remove
`internalArgTestOnlyFeature` test