squash! Drop language version 1.7
Looks like the number of Kotlin 1.7 users is roughly the same as
the number of users of 1.6. If we feel this number of users is safe for
dropping 1.6, then it's probably also safe for 1.7.

`fixCompilerArgs` is needed to account for the use of
`-XXLanguage:-TypeEnhancementImprovementsInStrictMode` in `kotlin-dsl`.
Because 1.7 is dropped, disabling the feature becomes prohibited. This
will be fixed in Gradle 9, but until then this code is required.

For example, try compiling
`:gradle-settings-conventions:jvm-toolchain-provisioning:compileKotlin`
with the bootstrap compiler.

^KT-71793