[FIR] Move CheckArguments stage before CheckContextArguments
Before this change, arguments didn't contribute any constraints to a
candidate's system before we checked context arguments.
In the test genericContextAmbiguity.kt we have potential context
arguments `Describer<A>` and `Describer<B>`, and both were considered
applicable because we didn't have any constraints on T, yet.
After this change, we can filter out `Describer<B>` because it violates
the constraint `A <: T`.

After the change, we need to skip processing of "fake" context arguments
in invoke convention candidates in CheckArguments stage
because they aren't replaced with real expressions, yet.
However, the same processing also happens in CheckContextArguments,
therefore, it was redundant to begin with.

#KT-75150 Fixed