[Analysis API] Ignore context elements of unsupported kinds
For injected files, the 'getContext()' points to the containing file's
element. The element, together with the context file, is not required
to be the Kotlin file. In the commit, such unsupported context elements
are filtered out.

In theory, Java files can also be supported in the check. However, it
seems to be an unusual scenario, so it's safer not to support it until
really needed.