kotlinx-metadata: remove builtinsAnnotations
Similarly to klib-specific annotations, use the common annotations
fields added in KT-57919.

Also fix the missing logic for reading/writing property accessor
annotations.

Note that kotlinp output for optional annotations in case of K1 has
changed. This is because optional annotations are not normal classes,
their metadata is stored inside the JVM module metadata. So when
JvmKotlinp rendered them, it tried to lookup annotations according to
JVM metadata rules, which is only in the new place (the
`KmClass.annotations` field). However, K1 writes them only to the old
place (builtins annotations extension), so JvmKotlinp could not find any
annotations, in particular `@OptionalExpectation`. Now all annotations
are always present in `KmClass.annotations` and there are no
platform-specific ways to load annotations, so `@OptionalExpectation`
is found and rendered.