FIR: do not write annotations on optional annotation classes twice
Since optional annotation classes are not materialized on JVM, we store
them in a special proto field `optional_annotation_class` in the module
(see `jvm_module.proto`). Annotations on these classes (note that there
must be at least one annotation `@OptionalExpectation`) were stored, a
bit confusingly, into an extension `BuiltInsProtoBuf.classAnnotation`,
which is normally used for classes compiled to `.kotlin_builtins` and
`.kotlin_metadata`, NOT for anything in JVM. It was done like this
because there was no other place to store annotations at that time.

After KT-57919 was implemented, the code in `FirJvmBackendExtension`
started to write these annotations not only on the extension, but also
on the new field `ProtoBuf.Class.annotation`. This happened because this
code uses `FirJvmSerializerExtension`, which is used for normal JVM
class files, and where a part of logic of serializing classes is
now serializing their annotations to the new field.

This led to annotations being serialized twice. It wasn't affecting
anything because the reading code in `OptionalAnnotationClassesProvider`
was only loading them from the old place
(`BuiltInsProtoBuf.classAnnotation`). However, after certain further
changes in kotlinx-metadata related to KT-75290, we'd have duplicates of
annotations provided via `KmClass.annotations`.

The solution is to use only the old place to store annotations before
2.2, and use only the new place starting from 2.2. The reading code is
adapted to read annotations from both places (it doesn't need to know
the version because we can check if one of the annotation lists is not
empty).