FIR: Introduce FirSuperReceiverExpression
Previously, `super<.>` expressions were represented
as FirPropertyAccessExpression, which looks both unnatural, but also
it's the only case when FirPropertyAccessExpression.calleeReference
is not FirNamedReference.

This refactoring is expected to help for context-sensitive resolution
once all property access would be built upon named references.

NB: This commit doesn't leave the build green without some of the
further commits

^KT-75315 Related