[Compose] Compose K1 compatibility IR linkage workaround
When used from K2, the already published K1 Compose library may have
a cross-module `IrGetField` access to a "stability" field that became
private in K2. This causes linkage errors (either on the partial
linkage level, or later on the Native binaries linkage level).

This workaround is applied during the partial linkage phase. It replaces
such field access by an invocation of the appropriate "stability" getter
function.