[FIR] Optimize flexible types that have trivial bounds
We introduce the property ConeFlexibleType.isTrivial that signifies
that the upper bound is a trivial nullable copy of the lower bound.
When recursively descending into flexible types (e.g., in substitution,
enhancement, utils like contains, ...), we can skip half of the
computations on trivial flexible types.

This fixes exponential performance on deeply nested trivial flexible
types.

#KT-74824 Fixed