amend! [FIR] FirExpectActualResolver Refactor. Split single `filter` call into two appropriate calls
[FIR] FirExpectActualResolver Refactor. Split single `filter` call into two appropriate calls

There are two branches:
1. Matching callables inside classes
2. Matching top-level callables

In the previous boolean expression:

    (expectContainingClass != null && expectSymbol.visibility != Visibilities.Private /*match non-private fake overrides*/ ||
            expectSymbol.isExpect && expectSymbol.moduleData in transitiveDependsOn)

The first part of boolean "or" refers to the the first branch, the
second part refers to the second branch.

If you wonder where does
`!expectSymbol.isFakeOverride(expectContainingClass) ||` come from.
It is needed to keep `compiler/testData/diagnostics/tests/multiplatform/headerClass/privateMembers.kt` green.
An allowance for non fake-overrides implicitly contained in the second
part of the combined boolean "or".

Motivation for the refactor: it is a preparation for KT-73557