Inference: Report missing type mismatch related to fork points
Previously, if there was some fork point, we considered the whole
subtyping as successful independently of components' subtyping results.

Probably, it was made like this because we assumed that the error
would be anyway reported once we process the fork point.

But the problem with this approach is that we only record
_type variable_ constraints for fork points, while there might be
inconsistencies within _initial_ constraints in another part
of the types irrelevant to any type variable.

For example, in failForkPointWhenNoneOfThemSuccessful.kt,
we've got
Base<*, Token.TokenB, Data> & Base<Type<out Data>, Token.TokenB, Data>
    <: Base<Av, Token.TokenA, Kv>

So, both branches contradict in the second argument
(Token.TokenB !<: Token.TokenA), but we just ignored that fact
and gather variables' constraints for the branches.

^KT-75444 Fixed