[K/N] Add a comment to a hack in `convertBlockPtrToKotlinFunction` impl
fd86955 added a new intrinsic -- `convertBlockPtrToKotlinFunction`.
The implementation contains a hack -- it relies on
`ObjCBlockPointerValuePassing.cToBridged` being a no-op.

It is possible to fix this, but that would require much more effort,
unjustified by the role of the intrinsic.

So, this commit adds a detailed comment instead to explain the hack and
outline the alternatives.

^KT-75207 Fixed