K2: add test ensuring KT-75649 case works without NPE in master
The case with lateinit initialized with "null of platform type"
in init {} block throws NPE in 1.9,
but unfortunately does not do it in both 2.0 and 2.1.
Because of that we had to keep it working in 2.1.20 and,
most probably, will need a deprecation cycle or at least
some warning to forbid this case later.