FIR2IR: partial revert of KT-71752 fix for lateinit lvar case
During the fix of KT-71752, we added not-null runtime checks
during assignments of platform type values to not-null variables/properties.
These checks existed in K1 but were forgotten in K2. However,
this is an additional runtime exception in a minor release,
and it looks like that we have a code in the wild using
not-null lateinits on left sides and platform nulls on right sides.
For this reason, it was decided to delay introduction of these checks
to a major (2.2.0) release (KT-75677), with possible introduction
of a deprecation warning in some patch release (KT-75676).

#KT-75649 Fixed