[Repl] Clear K2 Repl message collector before compiling each snippet
This change ensures that the K2 and K1 Repl behave similar in terms of what messages are being reported, i.e. only messages related to the current snippet being compiled. Before this change, K2 would collect all messages seen for the entire lifetime of the Repl session and then dump all of them when an error occurred. This could lead to duplicate or confusing messages being reported to the user, e.g. DEBUG level messages from an earlier snippet.

A part of ^KT-74685