[Analysis API] Provide support for modules targeting `JS` + `WASM/JS`
It is a case when a module has several different target platforms, however, is not a common module.
When some module targets `JS` and `WASM/JS` at the same time, the required logic is identical to the one used for a single `JS` target platform.
That's because a set of session platform components used for `WASM/JS` is a subset of `JS` platform modules.

Additionally, `JsPlatformAnalyzerServices` and `WasmPlatformAnalyzerServices` are identical to each other.

^KT-72047 fixed